+++ Todo-list/index.html
@@ -11,7 +11,7 @@
 
 <body>
 	<div id="root"></div>
-	<script src="./main.js"></script>
+	<script src="./src/index.js" type="module"></script>
 </body>
 
 </html>
