+++ Todo-list/src/utils/createElement.js
@@ -1,0 +1,7 @@
+export const createElement = (tag, options) => {
+	const elem = document.createElement(tag)
+	Object.keys(options).forEach(item => {
+		elem[item] = options[item]
+	})
+	return elem
+}
